Added try/catch around IndexOutOfBoundsExceptions
Approximately 3% of the tested 2018 Retrosheet game accounts threw IndexOutOfBoundsExceptions originating in BoxscoreGameAccount.readPlay. These exceptions are thrown when a runner is not on the base a that a file play line says that he is. Rather than using a heuristic to handle these cases, those games are catalogued and ignored so that the program can continue to run as normal, effectively jumping over those excepted games. At the end of the program's run, those recorded exceptions are printed to the console.